use tokenSets throughout code
