+++ packages/cli/src/tokens/types.ts
@@ -1,6 +1,6 @@
 import type { CssColor } from '../colors/types.js';
 
-export type Token = { $value: string; $type: string };
+export type Token = { $value: string | Record<string, string>[]; $type: string };
 export type TokensSet = {
   [key: string]: Token | TokensSet;
 };
@@ -12,35 +12,6 @@ export type ColorModeTokens = {
 
 // Define types for typography tokens
 export type TypographyTokens = TokensSet;
-
-// Define the main Tokens type
-export type Tokens = {
-  primitives: {
-    'colors-scheme': {
-      light: ColorModeTokens;
-      dark: ColorModeTokens;
-      contrast?: ColorModeTokens;
-    };
-    typography: {
-      primary: TokensSet;
-      secondary: TokensSet;
-    };
-  };
-  semantic: {
-    modes: {
-      'main-color': Record<string, TokensSet>;
-      'support-color': Record<string, TokensSet>;
-    };
-    color: TokensSet;
-    style?: TokensSet;
-  };
-  themes: Record<string, TokensSet>;
-};
-
-export type SemanticModes = {
-  'main-color': Record<string, TokensSet>;
-  'support-color': Record<string, TokensSet>;
-};
 
 export type Colors = {
   main: Record<string, CssColor>;
