+++ packages/cli/src/tokens/write.ts
@@ -2,9 +2,8 @@ import path from 'node:path';
 import type { ThemeObject } from '@tokens-studio/types';
 import chalk from 'chalk';
 import * as R from 'ramda';
-import type { ColorScheme } from '../colors/types.js';
-import { cp, mkdir, readFile, writeFile } from '../utils.js';
-import type { Collection, File, Theme, Tokens, TokensSet, TypographyModes } from './types.js';
+import { mkdir, readFile, writeFile } from '../utils.js';
+import type { Theme, TokensSet } from './types.js';
 import { generateMetadataJson } from './write/generate$metadata.js';
 import { generateThemesJson } from './write/generate$themes.js';
 
@@ -12,42 +11,19 @@ const DIRNAME: string = import.meta.dirname || __dirname;
 
 export const stringify = (data: unknown) => JSON.stringify(data, null, 2);
 
-const generateColorSchemeFile = (scheme: ColorScheme, name: Collection, tokens: TokensSet, outPath: string): File => {
-  const path = `${outPath}/primitives/modes/color-scheme/${scheme}`;
-  return {
-    data: stringify(tokens),
-    dir: path,
-    path: `${path}/${name}.json`,
-  };
-};
-
-const generateTypographyFile = (
-  folder: TypographyModes,
-  name: Collection,
-  tokens: TokensSet,
-  outPath: string,
-): File => {
-  const path = `${outPath}/primitives/modes/typography/${folder}`;
-  return {
-    data: stringify(tokens),
-    dir: path,
-    path: `${path}/${name}.json`,
-  };
-};
-
 type WriteTokensOptions = {
   outDir: string;
-  tokens: Tokens;
   theme: Theme;
   /** Dry run, no files will be written */
   dry?: boolean;
+  tokenSets: Map<string, TokensSet>;
 };
 
 export const writeTokens = async (options: WriteTokensOptions) => {
   const {
     outDir,
-    tokens,
-    theme: { name: themeName, colors, borderRadius },
+    tokenSets,
+    theme: { name: themeName, colors },
     dry,
   } = options;
   const targetDir = path.resolve(process.cwd(), String(outDir));
@@ -56,9 +32,6 @@ export const writeTokens = async (options: WriteTokensOptions) => {
   let themes = [themeName];
 
   await mkdir(targetDir, dry);
-
-  // Copy default files
-  await cp(path.join(DIRNAME, './template/design-tokens'), targetDir, dry, (src) => !src.includes('template.json'));
 
   try {
     // Update with existing themes
@@ -83,47 +56,13 @@ export const writeTokens = async (options: WriteTokensOptions) => {
   await writeFile($themesPath, stringify($theme), dry);
   await writeFile($metadataPath, stringify($metadata), dry);
 
-  /*
-   * Colors
-   */
+  for (const [set, tokens] of tokenSets) {
+    // Remove last part of the path to get the directory
+    const dirPath = path.join(targetDir, R.init(R.split('/', set)).join('/'));
+    await mkdir(dirPath, dry);
 
-  // Create main-color and support-color modes for the custom colors
-
-  for (const mode of Object.entries(tokens.semantic?.modes || {})) {
-    const [category, colors] = mode;
-    const categoryPath = path.join(targetDir, 'semantic', 'modes', category);
-    await mkdir(categoryPath, dry);
-
-    for (const [colorName, color] of Object.entries(colors)) {
-      await writeFile(path.join(categoryPath, `${colorName}.json`), stringify(color), dry);
-    }
-  }
-
-  // Create semantic color file
-  await writeFile(path.join(targetDir, `semantic/color.json`), stringify(tokens.semantic.color || {}), dry);
-
-  // Create theme file
-  await mkdir(path.join(targetDir, 'themes'), dry);
-  await writeFile(path.join(targetDir, `themes/${themeName}.json`), stringify(tokens.themes[themeName] || {}), dry);
-
-  // Create color scheme and typography modes
-  const colorScheme = tokens.primitives['colors-scheme'];
-  const typography = tokens.primitives.typography;
-
-  const files: File[] = [
-    generateColorSchemeFile('light', themeName, colorScheme.light[themeName], targetDir),
-    generateColorSchemeFile('light', 'global', colorScheme.light.global, targetDir),
-    generateColorSchemeFile('dark', themeName, colorScheme.dark[themeName], targetDir),
-    generateColorSchemeFile('dark', 'global', colorScheme.dark.global, targetDir),
-    // generateColorSchemeFile('contrast', themeName, colorScheme.contrast[themeName], targetDir),
-    // generateColorSchemeFile('contrast', 'global', colorScheme.contrast.global, targetDir),
-    generateTypographyFile('primary', themeName, typography.primary, targetDir),
-    generateTypographyFile('secondary', themeName, typography.secondary, targetDir),
-  ];
-
-  for (const file of files) {
-    await mkdir(path.resolve(file.dir), dry);
-    await writeFile(path.resolve(file.path), file.data, dry);
+    const filePath = path.join(targetDir, `${set}.json`);
+    await writeFile(filePath, stringify(tokens), dry);
   }
 
   console.log(
